Second round of type checks. This was a more tricky one, e.g.,  - for std::string, length() returns size_t (basically unsigned int), but  - for QString, length() returns int
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@10721 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/base/MidiDevice.cpp
@@ -207,7 +207,7 @@ MidiDevice::generateDefaultControllers()
         { "PitchBend", Rosegarden::PitchBend::EventType, "<none>", "0", "16383", "8192", "1", "4", "-1" }
     };
 
-    for (unsigned int i = 0; i < sizeof(controls) / sizeof(controls[0]); ++i) {
+    for (size_t i = 0; i < sizeof(controls) / sizeof(controls[0]); ++i) {
 
         Rosegarden::ControlParameter con(controls[i][0],
                                          controls[i][1],
