+++ test/base/realtime.cpp
@@ -25,7 +25,7 @@ int main(int argc, char **argv)
 
     int good = 0, bad = 0;
 
-    for (int i = 0; i < sizeof(rates)/sizeof(rates[0]); ++i) {
+    for (size_t i = 0; i < sizeof(rates)/sizeof(rates[0]); ++i) {
 
 	int rate = rates[i];
